requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.0
six==1.16.0
slack-sdk==3.15.1
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.0.0
yarl==1.7.2
